// web/src/types/progressMilestone.ts
export type ProgressMilestone = {
  id: string
  publicCode: string
  authorName: string
  courseName: string
  watchedLessonsCount: number
  totalLessonsCount: number
  completionPercentage: number
  courseStartedAt: string
  courseFinishedAt: string
  generatedAt: string
  shareUrl: string
  imageUrl: string
}